AI-Chatbot/backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.3.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
bcrypt==4.1.3
pyjwt==2.8.0
httpx==0.27.0

# AI
langchain==0.2.6
langchain-openai==0.1.14
langchain-chroma==0.1.2
langchain-community==0.2.6
langchain-experimental==0.0.62
langchain-text-splitters==0.2.2
openai==1.35.7
chromadb==0.5.3

# File handling
aiofiles==23.2.1
python-multipart==0.0.9
python-magic==0.4.27

# Document loaders
pypdf==4.2.0
unstructured[xlsx]==0.14.5

# Google Sheets
gspread==6.1.2
google-auth==2.31.0

# Data
pandas==2.2.2

# Dev / testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
python-dotenv==1.0.1
